Release/2.2.2

+++ django-os2webscanner/os2webscanner/models/statistic_model.py
@@ -30,11 +30,11 @@ class Statistic(models.Model):
     files_is_dir_count = models.IntegerField(default=0)
 
     # Information from the prescanner
-    supported_size = models.IntegerField(default=0)
+    supported_size = models.BigIntegerField(default=0)
     supported_count = models.IntegerField(default=0)
-    relevant_size = models.IntegerField(default=0)
+    relevant_size = models.BigIntegerField(default=0)
     relevant_count = models.IntegerField(default=0)
-    relevant_unsupported_size = models.IntegerField(default=0)
+    relevant_unsupported_size = models.BigIntegerField(default=0)
     relevant_unsupported_count = models.IntegerField(default=0)
